edit 2-calloc.c

+++ 0x0C-more_malloc_free/2-calloc.c
@@ -7,7 +7,7 @@
 */
 void *_calloc(unsigned int nmemb, unsigned int size)
 {
-	unsigned int i = 0;
+	unsigned long long i = 0;
 	char *p = malloc(sizeof(char) * nmemb * size);
 
 	if (nmemb == 0 || size == 0 || p == 0)
